clean up and separate workers
