+++ deploy/clowdapp.yaml
@@ -3579,8 +3579,8 @@ objects:
               fieldPath: metadata.name
         - name: DEVELOPMENT
           value: ${DEVELOPMENT}
-        - name: ENABLE_SUBS_EXTRACTION_DEBUG
-          value: ${ENABLE_SUBS_EXTRACTION_DEBUG}
+        - name: ENABLE_SUBS_DEBUG
+          value: ${ENABLE_SUBS_DEBUG}
         - name: CELERY_LOG_LEVEL
           value: ${CELERY_LOG_LEVEL}
         - name: KOKU_LOG_LEVEL
@@ -4675,5 +4675,5 @@ parameters:
   required: true
   value: subs_extraction
 - displayName: SUBS Extraction debug
-  name: ENABLE_SUBS_EXTRACTION_DEBUG
+  name: ENABLE_SUBS_DEBUG
   value: "False"
