+++ deploy/kustomize/patches/worker-subs-extraction.yaml
@@ -39,8 +39,8 @@
               fieldPath: metadata.name
         - name: DEVELOPMENT
           value: ${DEVELOPMENT}
-        - name: ENABLE_SUBS_EXTRACTION_DEBUG
-          value: ${ENABLE_SUBS_EXTRACTION_DEBUG}
+        - name: ENABLE_SUBS_DEBUG
+          value: ${ENABLE_SUBS_DEBUG}
         - name: CELERY_LOG_LEVEL
           value: ${CELERY_LOG_LEVEL}
         - name: KOKU_LOG_LEVEL
@@ -190,5 +190,5 @@
   path: /parameters/-
   value:
     displayName: SUBS Extraction debug
-    name: ENABLE_SUBS_EXTRACTION_DEBUG
+    name: ENABLE_SUBS_DEBUG
     value: "False"
