+++ koku/subs/__init__.py
@@ -2,5 +2,3 @@
 #
 # Copyright 2023 Red Hat Inc.
 # SPDX-License-Identifier: Apache-2.0
-#
-"""Subscription Watch Data Extraction"""
