+++ koku/subs/extraction/__init__.py
@@ -1,0 +1,6 @@
+# noqa
+#
+# Copyright 2023 Red Hat Inc.
+# SPDX-License-Identifier: Apache-2.0
+#
+"""Subscription Watch Data Extraction Worker"""
